+++ backend/src/geo-actions/geo-actions.module.ts
@@ -1,0 +1,11 @@
+import { Module } from '@nestjs/common';
+import { FirebaseModule } from 'src/firebase/firebase.module';
+import { GeoActionsController } from './geo-actions.controller';
+import { GeoActionsService } from './geo-actions.service';
+
+@Module({
+  imports: [FirebaseModule],
+  controllers: [GeoActionsController],
+  providers: [GeoActionsService],
+})
+export class GeoActionsModule {}